Rest API enabled with flask-jwt-extended

+++ flask_rest/resources/item.py
@@ -1,9 +1,12 @@
 from flask import Flask,request,jsonify
-from flask_restful import Api,Resource,reqparse
-from flask_jwt import JWT,jwt_required
+from flask_restful import Resource,reqparse
+from flask_jwt_extended import (jwt_required,
+                                fresh_jwt_required,
+                                get_jwt_claims,
+                                jwt_optional,
+                                get_jwt_identity)
 from models.item import ItemModel
 
-from security import authenticate,identity
 
 class Item(Resource):
 
@@ -17,13 +20,14 @@ class Item(Resource):
                         required=True,
                         help="Every Item need a store id")
 
-    @jwt_required()
+    @jwt_required
     def get(self,name):
         item=ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {"status": 404,"msg": "{} not found".format(name)}
 
+    @fresh_jwt_required
     def post(self,name):
         data = self.parser.parse_args()
         item = ItemModel.query.filter_by(name=name,store_id=data["store_id"])
@@ -37,8 +41,12 @@ class Item(Resource):
                 return {"msg": "Error occured while inserting"}, 500
             return {"status": 201, "item": newitem.json()}
 
-
+    @jwt_required
     def delete(self,name):
+        claims=get_jwt_claims()
+        print(claims)
+        if not claims['is_admin']:
+            return {'message':'Admin privilage required'},401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete()
@@ -58,5 +66,11 @@ class Item(Resource):
         return msg
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"items":[ x.json() for x in ItemModel.query.all()]}
+        user_id=get_jwt_identity()
+        items=[ item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {"items":items},200
+        return {'items':[item['name'] for item in items],
+                'message':'More data available if you login'},200
